Removed resourceregistrycontroller, resourceregistrycontrollertest and moved the getresources to resourcecontroller, resourcecontrollertest
