+++ test/Altinn.AccessManagement.Tests/Controllers/ResourceControllerTest.cs
@@ -9,8 +9,10 @@ using System.Threading.Tasks;
 using Altinn.AccessManagement.Controllers;
 using Altinn.AccessManagement.Core.Clients.Interfaces;
 using Altinn.AccessManagement.Core.Models;
+using Altinn.AccessManagement.Core.Models.ResourceRegistry;
 using Altinn.AccessManagement.Core.Repositories.Interfaces;
 using Altinn.AccessManagement.Core.Services.Interfaces;
+using Altinn.AccessManagement.Models;
 using Altinn.AccessManagement.Tests.Mocks;
 using Altinn.AccessManagement.Tests.Util;
 using Altinn.AccessManagement.Tests.Utils;
@@ -213,6 +215,37 @@ namespace Altinn.AccessManagement.Tests.Controllers
             Assert.Equal(expected, actual);
         }
 
+        /// <summary>
+        /// Test case: GetResources returns a list of resources 
+        /// Expected: GetResources returns a list of resources filtered by resourcetype
+        /// </summary>
+        [Fact]
+        public async Task GetResources_valid_resourcetype()
+        {
+            // Arrange
+            List<ServiceResourceExternal> expectedResources = GetExpectedResources(ResourceType.MaskinportenSchema);
+
+            // Act
+            HttpResponseMessage response = await _client.GetAsync($"accessmanagement/api/v1/52004219/resources/maskinportenschema");
+            string responseContent = await response.Content.ReadAsStringAsync();
+            var options = new JsonSerializerOptions
+            {
+                PropertyNameCaseInsensitive = true,
+            };
+            List<ServiceResourceExternal> actualResources = JsonSerializer.Deserialize<List<ServiceResourceExternal>>(responseContent, options);
+
+            // Assert
+            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
+            AssertionUtil.AssertCollections(expectedResources, actualResources, AssertionUtil.AssertResourceExternalEqual);
+        }
+
+        private List<ServiceResourceExternal> GetExpectedResources(ResourceType resourceType)
+        {
+            List<ServiceResourceExternal> resources = new List<ServiceResourceExternal>();
+            resources = TestDataUtil.GetResources(resourceType);
+            return resources;
+        }
+
         private HttpClient GetTestClient()
         {
             HttpClient client = _factory.WithWebHostBuilder(builder =>
@@ -221,6 +254,7 @@ namespace Altinn.AccessManagement.Tests.Controllers
                 {
                     services.AddSingleton<IResourceMetadataRepository, ResourceMetadataRepositoryMock>();
                     services.AddSingleton<IPostConfigureOptions<JwtCookieOptions>, JwtCookiePostConfigureOptionsStub>();
+                    services.AddSingleton<IResourceRegistryClient, ResourceRegistryClientMock>();
                 });
             }).CreateClient(new WebApplicationFactoryClientOptions { AllowAutoRedirect = false });
